med/pyproject.toml
[tool.poetry]
name = "qa"
version = "0.1.0"
description = ""
authors = ["Leonid Kuligin <kuligin@google.com>", "Aleksandr Ostapenko <aostapenko@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.8"
langchain-google-community = "^2.0.2"
langchain-google-vertexai = "^2.0.20"
pydantic = ">=2.7,<2.10"
langgraph = "^0.3.31"
grpcio = "^1.68.0"
langchain-mistralai = "^0.2.2"
anthropic = {extras = ["vertexai"], version = "^0.39.0"}
beautifulsoup4 = "^4.12.3"
langchain-openai = "^0.3.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
